requirements.txt
aiomysql==0.2.0
databases==0.9.0
fastapi==0.111.1
nltk==3.8.1
numpy<2.0
pydantic==2.8.2
python-dotenv==1.0.1
scikit-learn==1.5.1
sentence-transformers==3.0.1
sqlalchemy==2.0.32
torch==2.2.2
transformers==4.43.3
uvicorn==0.30.4